feat: añadir comprobación usuario admin backend

+++ backend/app/routes/product_routes.py
@@ -1,8 +1,10 @@
 from flask import Blueprint, jsonify, request
+from flask_jwt_extended import jwt_required, get_jwt
 
 from app.models.exceptions import ErrorInternoException, ProductoNoEncontradoException, CampoIncorrectoException
 from app.models.producto import Producto
 from app.services.product_services import obtener_todos_los_productos, obtener_producto_id, insertar_producto_base, eliminar_producto_base, actualizar_datos_producto
+from app.services.auth_services import comprobar_usuario_es_admin
 
 producto_bp = Blueprint("productos", __name__)
 
@@ -28,7 +30,13 @@ def get_producto_id(id):
 
 # POST INSERTAR PRODUCTO
 @producto_bp.route("/", methods=["POST"])
+@jwt_required()
 def post_insertar_producto():
+    data = get_jwt()
+
+    if comprobar_usuario_es_admin(data) is False:
+        return jsonify({"error": "Acción cancelada. El usuario no es admin"}), 403
+
     data = request.get_json()
 
     if not data or not data.get("nombre") or not data.get("descripcion") or (data.get("stock") is None) or (data.get("precio") is None) or not data.get("url") or not data.get("categoria"):
@@ -49,7 +57,13 @@ def post_insertar_producto():
 
 # DELETE ELIMINAR PRODUCTO
 @producto_bp.route("/<int:id>", methods=["DELETE"])
+@jwt_required()
 def delete_producto(id):
+    data = get_jwt()
+
+    if comprobar_usuario_es_admin(data) is False:
+        return jsonify({"error": "Acción cancelada. El usuario no es admin"}), 403
+
     try:
         eliminar_producto_base(id)
 
@@ -61,7 +75,13 @@ def delete_producto(id):
 
 # ACTUALIZAR PRODUCTO
 @producto_bp.route("/<int:id>", methods=["PUT"])
+@jwt_required()
 def actualizar_producto(id):
+    data = get_jwt()
+
+    if comprobar_usuario_es_admin(data) is False:
+        return jsonify({"error": "Acción cancelada. El usuario no es admin"}), 403
+
     data = request.get_json()
 
     if not data or not data.get("nombre") or not data.get("descripcion") or (data.get("stock") is None) or (data.get("precio") is None) or not data.get("url") or not data.get("categoria"):
